requirements.txt
# Brain Tumor Classification - Requirements for Streamlit Cloud Deployment
# Python 3.10+ recommended

# Core Framework
streamlit>=1.30.0

# Deep Learning
tensorflow>=2.15.0
keras>=3.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Image Processing
Pillow>=10.0.0

# ML Utilities
scikit-learn>=1.3.0

# Optional (for development)
# jupyter>=1.0.0
# ipython>=8.0.0
